Weekly Challenge # 359 - Alteryx Academy
Almost any data project will include dates. Working with dates is a common task and is important for downstream tasks such as reporting. In this week's challenge we are going to cover a common date manipulation that many will encounter. 

In this scenario, we have a column of date data that is in the format of MM/dd/yyyy. We wantto reformat that data to yyyyquarter. Your goal for this challenge is to convert the date format from MM/dd/yyyy to yyyyquarter. As an example, change 2/15/2023 to 2023Q1.